clean up/add comments to NASA translate flow

+++ data-engineering-resources/NASA_ch10_flows/translate_NASA_1553_flow.py
@@ -16,7 +16,8 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
 
     n = metaflow.Parameter(
         "n",
-        help="Number of Parsed 1553 items to translate and upload.",
+        help="Number of ch10_parsed baskets to translate and upload. "
+             "Default is to translate all parsed baskets.",
         required=False,
         default=None,
         type=int
@@ -30,6 +31,7 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
     )
 
     def extract_metadata(self):
+        '''Gather metadata from translation.'''
         translate_metadata = {}
         meta_file = os.path.join(self.local_translate_path, "parsed_data_translated", "_metadata.yaml")
         
@@ -41,12 +43,104 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
             translate_metadata['translate_metadata'] = metadata
             
         return translate_metadata
+    
+    def translate_basket(self, basket):
+        '''Translate parsed 1553 data and extract metadata.
+        
+        Given a parsed 1553 basket, check that there exists only one 1553.parquet
+        in that basket. Then download the basket to a temporary directory, and translate
+        said parsed data. Extract and return the generated metadata.
+        
+        Parameters
+        ----------
+        basket (str): path to a basket of type ch10_parsed in s3.
+        
+        Returns
+        -------
+        translate_metadata (dict): metadata generated during tip_translate.
+        '''
+        opal_s3fs = s3fs.S3FileSystem(client_kwargs = {'endpoint_url': os.environ['S3_ENDPOINT']})
+        
+        basket_contents = opal_s3fs.ls(basket)
+
+        #check that there is one parsed path and get the path to it
+        s3_parsed_path = [x for x in basket_contents if '1553' in x and x.endswith('.parquet')]
+        if len(s3_parsed_path) != 1: 
+            raise Exception(f'Basket does not have parsed data, skipping: {basket}')
+
+        s3_parsed_path = s3_parsed_path[0]
+
+        self.local_parsed_dir = os.path.join(self.local_dir_path, 'parsed_data.parquet')
+        os.mkdir(self.local_parsed_dir)
+
+        opal_s3fs.get(s3_parsed_path, self.local_parsed_dir, recursive = True)
+        self.local_translate_path = os.path.join(self.local_dir_path, 'translated_output')
+        os.mkdir(self.local_translate_path)
+
+        # run tip translate
+        subprocess.run(
+            [
+                "tip_translate_1553",
+                self.local_parsed_dir,
+                self.local_dts_path,
+                "-L",
+                "off",
+                "--output_path",
+                self.local_translate_path
+            ]
+        )
+        
+        return self.extract_metadata()
+    
+    def upload_translate_basket(self, basket, translate_metadata):
+        '''Upload translated data to MinIO.
+        
+        From s3, get the name of the ch10 and uuid of the parent basket. Then 
+        Upload the basket with the below information.
+        
+        Parameters
+        ----------
+        basket (str): path to a basket of type ch10_parsed in s3.
+        translate_metadata (dict): all metadata generated during tip_translate.
+                             
+        Returns
+        -------
+        basket_upload_path (str): path to where the basket was uploaded,
+                                  returned from self.metaflow_upload_basket
+        '''  
+        opal_s3fs = s3fs.S3FileSystem(client_kwargs = {'endpoint_url': os.environ['S3_ENDPOINT']})
+        
+        manifest_data = {}
+        manifest_path = os.path.join(basket, 'basket_manifest.json')
+        with opal_s3fs.open(manifest_path, 'rb') as file:
+            manifest_data = json.load(file)
+
+        self.parsed_id = manifest_data['uuid']
+        parent_ids = [self.parsed_id, self.dts_id]
+
+        parsed_metadata = {}
+        metadata_path = os.path.join(basket, 'basket_metadata.json')
+        with opal_s3fs.open(metadata_path, 'rb') as file:
+            parsed_metadata = json.load(file)
+
+        ch10_name = parsed_metadata['ch10name']
+        
+        translate_metadata['ch10name'] = ch10_name
+
+        #build upload_dicts
+        upload_dicts = []
+        for f in os.scandir(self.local_translate_path):
+            upload_dicts.append({'path':f.path,'stub':False})
+
+        return self.metaflow_upload_basket( upload_dicts,
+                                            self.bucket_name,
+                                           'ch10_translated_1553',
+                                            label = ch10_name,
+                                            metadata = translate_metadata)
 
     @step
     def start(self):
-        """Create empty temporary directory for parsed and translated data.
-        
-        """
+        '''Create empty temporary directory for parsed and translated data.'''
         #create temporary directory to put data files locally
         self.local_dir = tempfile.TemporaryDirectory()
         self.local_dir_path = self.local_dir.name
@@ -57,7 +151,7 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
         
     @step
     def get_dts_file(self):
-        
+        '''Get latest DTS file from S3 and save locally for translation.'''
         opal_s3fs = s3fs.S3FileSystem(client_kwargs = {'endpoint_url': os.environ['S3_ENDPOINT']})
         
         index = create_index_from_s3(self.bucket_name, 'schema.json')
@@ -86,9 +180,9 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
         
     @step
     def translate_parsed_1553(self):
-        '''
-        '''
+        '''Translate ch10_parsed data from S3 and upload translated data as a ch10_translated_1553 basket'''
         opal_s3fs = s3fs.S3FileSystem(client_kwargs = {'endpoint_url': os.environ['S3_ENDPOINT']})
+        
         if not opal_s3fs.exists(self.bucket_name):
             raise FileNotFoundError(f"Specified Bucket Not Found: {self.bucket_name}")
         
@@ -97,75 +191,29 @@ class TranslateNASA1553Flow(opal.flow.OpalFlowSpec):
         self.ch10_parsed_baskets = ch10_index['address']
 
         if self.n is not None:
-            self.ch10_baskets = self.ch10_parsed_baskets[:self.n]
+            self.ch10_parsed_baskets = self.ch10_parsed_baskets[:self.n]
 
         num_baskets = len(self.ch10_parsed_baskets)
         count = 1
-        for basket in self.ch10_parsed_baskets:            
-            basket_contents = opal_s3fs.ls(basket)
-            
-            #check that there is one parsed path and get the path to it
-            s3_parsed_path = [x for x in basket_contents if '1553' in x and x.endswith('.parquet')]
-            if len(s3_parsed_path) != 1: 
-                print(f'Basket does not have parsed data, skipping: {basket}')
-                continue
+        for i, basket in enumerate(self.ch10_parsed_baskets):
+            try:
+                print(f'-- translating {i} of {num_baskets}: {basket}')
                 
-            s3_parsed_path = s3_parsed_path[0]
-            
-            local_parsed_dir = os.path.join(self.local_dir_path, 'parsed_data.parquet')
-            os.mkdir(local_parsed_dir)
-            
-            opal_s3fs.get(s3_parsed_path, local_parsed_dir, recursive = True)
-            self.local_translate_path = os.path.join(self.local_dir_path, 'translated_output')
-            os.mkdir(self.local_translate_path)
-            
-            manifest_data = {}
-            manifest_path = os.path.join(basket, 'basket_manifest.json')
-            with opal_s3fs.open(manifest_path, 'rb') as file:
-                manifest_data = json.load(file)
-                
-            self.parsed_id = manifest_data['uuid']
-            parent_ids = [self.parsed_id, self.dts_id]
-            
-            parsed_metadata = {}
-            metadata_path = os.path.join(basket, 'basket_metadata.json')
-            with opal_s3fs.open(metadata_path, 'rb') as file:
-                parsed_metadata = json.load(file)
-                
-            ch10_name = parsed_metadata['ch10name']
-            
-            print(f'-- translating {count} of {num_baskets}: {ch10_name}')
-            count += 1
-            
-            # run tip translate
-            subprocess.run(
-                [
-                    "tip_translate_1553",
-                    local_parsed_dir,
-                    self.local_dts_path,
-                    "-L",
-                    "off",
-                    "--output_path",
-                    self.local_translate_path
-                ]
-            )
-            
-            translate_metadata = self.extract_metadata()
-            translate_metadata['ch10name'] = ch10_name
+                translate_metadata = self.translate_basket(basket)
 
-            #build upload_dicts
-            upload_dicts = []
-            for f in os.scandir(self.local_translate_path):
-                upload_dicts.append({'path':f.path,'stub':False})
-            
-            self.metaflow_upload_basket(upload_dicts,
-                                        self.bucket_name,
-                                       'ch10_translated',
-                                        label = ch10_name,
-                                        metadata = translate_metadata)
+                basket_upload_path = self.upload_translate_basket(basket, translate_metadata)
+                        
+                print(f'basket successfully parsed and uploaded: {basket_upload_path}')
 
-            shutil.rmtree(self.local_translate_path)
-            shutil.rmtree(local_parsed_dir)
+            except Exception as e:
+                print(e)
+                print(f'translation failed: {basket})')
+                
+            finally:
+                if os.path.exists(self.local_translate_path):
+                    shutil.rmtree(self.local_translate_path)
+                if os.path.exists(self.local_parsed_dir):
+                    shutil.rmtree(self.local_parsed_dir)
 
         self.next(self.end)
